Add Events item to navigation menu.
Instead of doing this as part of the installation routine, we have
a persistent check on 'admin_init'. This ensures that the item is
added after the-events-calendar is activated. A flag is stored
in the database so that the item is not added multiple times (for
example, after the admin has deleted it).

Fixes #29.